Adjusted Date/Time Format Per Spec - DEV-1991
Adjusted the Date/Time Format per Activity Streams Specification:
https://www.w3.org/TR/activitystreams-core/#dates – The use of the
GMT/UTC offset is a valid alternative to the use of a static “Z”
character (“no offset”) placeholder, both according to the ISO
standards for dates and the AS spec itself. Given the international use
of these feeds, the inclusion of the GMT/UTC offset was deemed
appropriate and useful. Hopefully now the dates can be parsed easily
and per the spec.