Fixes "Missing" Objects in Activity Stream - DEV-1992
There was an issue with setting the database SELECT query offset when
paginating through the activity stream source table, and due to the
opaqueness of UUIDs it was not necessary apparent that it was
happening. This has now been fixed and the Object records which were
there all along, can now be seen in the Activity Stream. They were
hidden by earlier entries for Person/Group which happened to have been
transformed first. This update also adds sensible bounds to the limit
parameter, ensuring it is a positive integer, with a default of 100, a
minimum of 10, and a sensible maximum of 1000 items per page.