update native path and build and run scripts
I have moved the directory for native build and run scripts to a more suitable name and have also re-factored the dockerfile names and names of everything related including the container names and services to be more consistent.  I update the readme with instructions to run the db locally and the env requirements (since the db is not pulling it's env var's from vault) and to run things against staging.