Improve Title Classification - DEV-2242
Due to a lack of clarity in the schema documentation, primary titles
were previously classified_as as nothing more than ‘preferred’ but no
other classification was added - this update adds the AAT term
300417193 [Titles (General, Names)] to all titles emitted in the
JSON-LD to make it clear that a particular LinguisticObject is actually
a title, regardless of whether it is preferred or not. We should update
the schema documentation to recommend this change for clarity.