Map Museum Exhibitions - DEV-1718
Added transformer and mapping for Museum Exhibitions, modeled as a
PropositionalObject parent entity, and a variable number of related
Activity child entity records.

One incoming Exhibition record from the Musesum’s existing data set,
will result in multiple JSON-LD representations - one for the concept
of the Exhibition, and one for each of the instances of the Exhibition
being exhibited – modeled as the Activity of exhibiting the exhibition.
Slight changes to the BaseTransformer and BaseManager were needed to
enable this new use-case of out-of-bound storage of the generated
entities. See the related Jira tickets for more information.