Adjusted Object Number Mapping – DEV-2630
Adjusted `Object` identification number mappings; records will now only
contain mappings for both an accession number and display number if the
numbers are different; in cases where a valid display number is found,
it will be marked as the preferred term; in cases where no display
number is found, or if it is the same as the accession number, it will
not be mapped, and the accession number will be marked as preferred. In
cases where the `Object` being mapped is a Manuscript, its Manuscript
Department assigned number will also be mapped and classified
accordingly.