Evaluate ABCPLUS Dependency – DEV-2543
Reevaluated the ABC Plus dependency, and although the `@finalmethod`
decorator brought many advantages to ensure contract adherence, it was
recognized that the use of the library brings with it potential
long-term maintenance issues, as it was based on a fork of the standard
ABC library. As such the code has been updated to remove the dependency
on the ABC Plus library, and instead to continue making use of the core
features from the standard ABC library for its abstract class support.
The `@finalmethod` decorator and functionality has been replaced by
code comments noting to please not override the given methods, as doing
so could lead to unintended outcomes unless done with great care and
evaluation.