Additional Code Formatting Adjustments – DEV-3016
It is not clear why but a few of the recent code enhancements did not
get reformatted by _Black_ prior to being committed to the repository.
Although the formatting changes are minimal, it is best to resolve them
now before moving forward. The minor changes slightly affect the
formatting in four files.