Improve initial startup experience
To improve the experience getting the application up and runnning for
the first time, I've updated the README.md file to reflect what's needed to
start up the application.

This adds some default values to the .env.example file to minimize changes
needed to that file for first run.

I've also tweaked the docker-compose to make it so the database is saved
within the source directory, rather than using the system route--this
makes it easier to delete and retest the database initialization without
accidentally blowing away something else.

Finally, I removed the deprecated `link` properties in the docker-compose
configuration, since the application networking provides those by default.