Make application testable by using Flasks's App Factory Pattern
This commit modifies the structure of the application to use Flask's
App Factory Pattern, where the module generates an application, and a
small wsgi shim actually runs it. This will improve testability of the
Flask app, and follows the Flask documentation's pattern for testing.

Because this allows the application code to be treated as a module, it
required renaming the app from `app` to `flaskapp`, since the shared
code is also called `app`.  It's probably better to refactor the shared
code to have a clearer name, but that touches more files than I thought
wise in this commit.

Changes:
* Reduced the number of uniquely mounted volumes within docker compose
* Refactor the Dockerfile to treat /app as the workdir
	*  Added /app to PYTHON PATH (which might be a hack?  Not sure)
* Moved startup.py to flaskapp/__init.py__
* added blank __init.py__ files to make shared code and routes modules
* Changed the startup script to use wsgi.py
* Added sample Flask test