Consolidate Dockerfiles for web-service
This deletes the redundant Dockerfile and makes both the production
and development versions use the same Dockerfile.

It also transitions from an explicit PYTHONPATH declaration to using
the setup.py file to handle installing the application and locating
files within the path.  This follows by Flask documentation and Ben's
advice.

(I also removed the credit label, since it should now include Collin
and myself and that'd make it a bit long and not very useful.  We
can use this Git commit log to see who's worked on this internally.)