initialize SQLAlchemy Models
This creates two models that map closely to the existing DB schema.
There a couple of small changes: the biggest one is that the UUID column
is just a string, not an explicit UUID column.  This is a needed change
to accommodate record types that do not use UUIDs for their identifiers.

This also required moving some things around in the main __init__.py
file since the DB needed to exist within an app context.