Rewrite routes/records.py
This completely rewrites records.py to use SQLAlchemy instead of the
bespoke ORM. It merges the two routes into a single route using default
values, and it uses Flask's built-in config within the route rather than
relying on access to environment variables.

It also removes both the counter (since writing on every read can be a
performance hit) and the ETAG functionality, since Last-Modified and
ETAG are redundant.

This pull request also includes a test suite that is intended to be used
against a test database, since it rebuilds the DB on every test.

There is still a bug in the application--the datestamp for Last-Modified
is inaccurate (forced into GMT, but not actually GMT), but it will take
rewriting other parts of the code to correct.