Add new route for root OrderedCollection
This creates a specific route for /activity-streams, as well as a series
of tests for functionality.  This new route does not include any
non-standard properties, and takes its summary from a new ENV variable.

All DB access is through SQLAlchemy.

There's also an open question about sorting keys--the documentation
indicates there are cache considerations around key sorting, but it
makes it so much easier to read with the keys in the right order.

TODO: It does not yet handle multiple namespaces within a single app
correctly, in that is will return the total of all activities.