Giving up trying to fix git
Here's all the changes as a lump. Appending the previous commit messages
below.

Complete the rewrite of the web service

This commit finishes the rewrite of the application.  This version of
the application has 100% test coverage and no longer shares (almost) any
code with the Transformer (the camelcase code is the only exception.)

---

Add new route for root OrderedCollection

This creates a specific route for /activity-streams, as well as a series
of tests for functionality.  This new route does not include any
non-standard properties, and takes its summary from a new ENV variable.

All DB access is through SQLAlchemy.

There's also an open question about sorting keys--the documentation
indicates there are cache considerations around key sorting, but it
makes it so much easier to read with the keys in the right order.

TODO: It does not yet handle multiple namespaces within a single app
correctly, in that is will return the total of all activities.

---

Update circleci to use test db

---

Rewrite routes/records.py

This completely rewrites records.py to use SQLAlchemy instead of the
bespoke ORM. It merges the two routes into a single route using default
values, and it uses Flask's built-in config within the route rather than
relying on access to environment variables.

It also removes both the counter (since writing on every read can be a
performance hit) and the ETAG functionality, since Last-Modified and
ETAG are redundant.

This pull request also includes a test suite that is intended to be used
against a test database, since it rebuilds the DB on every test.

There is still a bug in the application--the datestamp for Last-Modified
is inaccurate (forced into GMT, but not actually GMT), but it will take
rewriting other parts of the code to correct.

---

Initialize SQLAlchemy Models

This creates two models that map closely to the existing DB schema.
There a couple of small changes: the biggest one is that the UUID column
is just a string, not an explicit UUID column.  This is a needed change
to accommodate record types that do not use UUIDs for their identifiers.

This also required moving some things around in the main __init__.py
file since the DB needed to exist within an app context.

---

Move DB instantiation and shared headers into helper module

---

install basic dependencies of SQLAlchemy

---

Remove sample test

This file was just used to verify the testing environment works,
with real tests it's no longer useful.

---

Break activityStream() into smaller functions

This pull request does not change functionality: it instead takes
sections of the larger activityStream() function and moves them
into other functions.

It also makes explicit the imports.

---

Create conftest.py file

This moves some of the more general fixtures into the standard location.