Upgraded CROM – DEV-3220
Upgraded CROM (`cromulent`) library to the latest commit
`589d28919f952c0620a04286a9716dc214dc5285`.

This also necessitated removing any references to the `ProperyInterest`
type which has been removed from the model due to the new rights
pattern being introduced by the SIG as per:
https://github.com/thegetty/crom/pull/102. Fortunately, we had not made
use of `PropetyInterest` thus far in the data mapping, so this was a
minor administrative change to remove any references from the library
imports, rather than a breaking change as it could be in some other
cases.