Make the base AS route calculate pages correctly
When updating the pagination code, I forgot to update the page counts
at the beginning and the end.

I also caught a bug that if there were no items, it would point to a
first and last page, but that page would 404.  Now those keys are not
exposed if published.