Fix People Entity Birth/Death TimeSpan End Dates – DEV-3372
Fixed `People` entity birth/death `TimeSpan` end dates to use the
correct `end_of_the_end` property for the end of the ending date rather
than `end_of_the_begin` as used unintentionally before.

Furthermore, the mapping has been modified to access the beginning and
ending dates of the date ranges of the birth and death activities,
which are padded appropriately as needed to correctly represent the
time span.