Update Person Entity Birth/Death Place Pattern – DEV-3373
Updated the `Person` entity birth/death place pattern to use the
`content` property of a `LinguisticObject` associated with the `Person`
entity via its `referred_to_by` property to store the birth/death place
descriptions, as at this time we do not have mappings to a controlled
vocabulary of birth/death places as implied by the use of the
`took_place_at` property used here previously.

Once we have completed a data reconciliation against a controlled
vocabulary of places, the use of `Place` entities and the
`took_place_at` property can be restored for birth/death places.