Adjust Artist/Maker Role Mapping – DEV-3107
Adjusted the artist/maker role mapping to utilize the production
technique used by the artist/maker as per the Linked.art schema versus
the production role. For example, instead of classifying the
artist/maker as a “photographer”, the schema calls for the process or
technique, “photography” in this case (or a specialization thereof), to
be used instead.