Update application to only include web service, not transformer
This also moves us to using the standard dockerfile location,
eliminating the duplicated dockerfiles, which was a problem for the
postgres container.  I've also removed the vault configuration from
the postgres system, since it's only a local development dependency.

Also updated requirements.txt to only include the libraries needed,
and moved some of the utilities into the base utility class.

Also pruned down the list of ENV variables.