Remove the optional namespacing, replacing it with mandatory namespaces
! This change renames the LOD_DEFAULT_URL_NAMESPACE env variable to be
APPLICATION_NAMESPACE, to avoid confusing people about what is and isn't
a default.

This change also modifies our testing methodology for local testing.  
We're not explicitly passing in `ns` as the APPLICATION_NAMESPACE value
for local testing.  We've also updated the CircleCI config to do the same.

We've also modified the way we load blueprints--the prefix is part of the
registration of the blueprint, not the blueprint code itself.  This
keeps all the namespace stuff in the `__init__.py` file. 

There are a couple minor tidying-ups that were done:

* making imports from standard libraries more explicit
* moving the utility tests out into their own file