update location directory name and modify for over-ride of entrypoint and command
moving the directory name to have just "docker" and not "docker-native" in the name.  Adjust the run command so that it has the over-ride for envconsul inclusion and the command over-ride also so it's incorporated to use envconsul.  This will make it consistent with how things run in staging and production.