Improve Neptune Exception Handling – DEV-4211
Added new tests for testing Neptune connectivity by utilising the
`requests_mock` library, which allows for HTTP requests made by the
`requests` library to be intercepted and responded to locally in code,
rather than over the network.

The new tests and the associated requests mocker allow for simulating a
successful connection to Neptune, as well as mimicking a failed
connection due to connectivity issues or time-outs.

This is particularly useful as our AWS Neptune instances are not
currently accessible from within CircleCI, so it is not currently
possible to perform actual testing against the Neptune instances during
the usual CI/CD process. Thus being able to mock the requests and
respond appropriately allows for the pass/fail code paths of the
Neptune integration to be more completely evaluated.

With the new tests in place via the support of the mocker, it is no
longer necessary to override the return values of the
`process_neptune_record_set()` or `graph_check_endpoint()` methods
using pytest’s `mocker.patch()` mechanism as all of the methods now
operate as though they have been able to communicate with Neptune, or
handle the raised `requests.exceptions.ConnectionError` appropriately
for the failure case.

An additional optional named parameter, `neptune_endpoint` has been
added to the `process_neptune_record_set()` method to ease configuring
the endpoint URL for testing with the appropriate mock URLs. If the
parameter is not specified, it falls back to the `NEPTUNE_ENDPOINT`
environment value as before, so no code changes are needed for the
usual runtime use case.