Add the docker authentication for parameter store
Need to add the docker authentication for the parameter store so that the envconsul private image can be pulled as part of the multi-stage build process.  Also, add other buildspec files for other branches so things stay in sync.  These will need to be merged to all other branches once in the master branch to keep things in sync.