Support Alternate Namespace for Neptune Loads – DEV-4527
After further research, we determined that configuring the namespace
was the desired approach versus configuring the base URL.

As such this branch has been updated to allow the application namespace
to be configured differently for Neptune loading for development or
specially staging loads versus allowing for the base URL to be
configured differently.

However the caveat here is that the corresponding development or
specially staged instances of the LOD Gateway should be hosted under
the same domain as their corresponding staging or production instances,
as only the namespace is now configurable differently for these special
use cases.