add yaml files for all branches to be included in master
People were checking in files directly to branches and things are out of sync.  If we want to move to the just master and prod branch model this needs to be clean first.  This should take care of it.  I am not going to re-name the container names in existing deployments as it's a lot of re-factoring and I have to rebuild the task defintions and buildspec's and ecr repos and pipelines.  We will leave the ones that are done for now and revisit that later.