Update JSON Schema & OpenAPI Documentation – DEV-5514
Updated JSON Schema and OpenAPI documentation of the LOD Gateway,
including documentation for the Activity Streams endpoints, the records
retrieval endpoint, and the credentialed `ingest` endpoint.

Removed dataset specific JSON Schemas from the repository as well as
references to the same from the OpenAPI specification that related to
the Museum Collection LOD data mapping project, as this documentation
is all now maintained within the Museum Collection LOD Transformer
repository instead along with the code that transforms the data.