Create Options for Recursive Prepending of Root URL & Namespace to Record ID Values – DEV-7034
This update brings an option to all LOD Gateway’s to control if the
prefixing of record IDs. This new functionality is controlled by the
optional `PREFIX_RECORD_IDS` environment variable. The default behavior
of recursively searching a record and prefixing any `”id”` values that
are not already HTTP(S) URLs, with the Gateway’s configured base URL
and namespace will remain in effect, if the `PREFIX_RECORD_IDS`
environment variable is absent, or if it is set to a value of
`RECURSIVE`. The new behaviors of only prefixing the top-level record
`”id”` if necessary is achieved by setting the environment variable to
a value of `TOP`, and to prevent any `”id”` values from being prefixed,
the environment variable may be set to a value of `NONE`.

The test suite has been updated with three new tests and some
additional sample data to test the three possible ID prefixing
behaviors, and the `README.md` has been updated with information on
configuring the desired behavior and the supported configuration
values. Furthermore, the update includes additional default values in
the `.env.exmaple` file to assist with installing and configuring a
Gateway on a local development system for the first time, by
demonstrating the configuration options in practice.