Additional Graph Expansion Exception Error Logging – DEV-7630
Added additional graph expansion exception error logging to hopefully
assist in debugging the graph expansion issues reported when loading
the staging instance of the Museum Collection LOD Gateway. The
additional logging will provide much more detail about the exception as
well as an export of the data that was being processed at the time of
the exception. Hopefully this will allow the cause to be traced back to
its source so that it can be reliably resolved. We can always remove
this additional logging if desired once the underlying issues are
resolved.