Additional Graph Expansion Exception Error Logging – DEV-7630
Removed requirement within `ingest.py`’s `graph_expand()` method for
JSON source documents to specify a `@context` attribute, as the
requirement was causing the test-suite to fail due to the exception
being raised when it was absent.

Should we try to expand a JSON document that lacks a `@context`
attribute, such as the JSON document strings specified in the test
suite?