Additional Graph Expansion Exception Error Logging – DEV-7630
Added the ability to capture the full stack trace from any call site,
so that we can get better insight into any non `JsonLdError` exceptions
being thrown during the call to `graph_expand()` due to the lack of
meaningful information being provided by the exception instances
themselves. Hopefully with the stack trace capture and logging now in
place we will finally be able to uncover the root-cause of the graph
expansion exceptions, and work to resolve them.