Additional Graph Expansion Exception Error Logging – DEV-7630
Fixed _Black_ formatting issues _again_! It is unclear why but _Black_
decided to change single-quotes to double quotes and add an arbitrary
space in a string slice, after taking a second pass at the code,
although it chose not to make these formatting changes on the first
pass! Clearly _Black_ is not as reliable or consistent as it could be!