Update README File for Open Sourcing – DEV-16094
Revised the `README` file for the LOD Gateway as part of the
preparations for releasing the LOD Gateway as Open Source Software.

Ensured that the `README` file is up-to-date, consistent, and is in
correspondence with the latest features of the Gateway, covering all
current functionality and configuration options, and that it provides
helpful references to newcomers or users who may be less familiar with
some of the foundational concepts, specifications, and/or frameworks
used to create the LOD Gateway. Added information about recently added
features, such as the dashboard, Gunicorn, and RDFLib. Removed
references to uWSGI.

Removed the `CHANGELOG` file as it was a remnant of the earlier LOD
Transformer project from which LOD Gateway was forked.

Created a logo for the project and added it to the repository and the
README file.

Updated the `.env.example` and `docker-compose.yml` to use the more
intuitive `fuseki` service name for the containerized Fuseki graph
store service rather than the prior `endpoint` name.

Updated the `docker-compose.yml` file allow the ports that the services
are exposed to the host to be configured, including Postgres and Fuseki
so that should a host already have services running on the default
ports, that they can be changed by configuration rather than having to
adjust the code.