LOD Gateway: Fix for POSTed SPARQL Queries – DEV-20233
Currently if a user tries to use a SPARQL client which makes HTTP POST
requests to LOD Gateway instances with the SPARQL query string
contained in the HTTP request body, and a `Content-Type` HTTP request
header of "application/sparql-query", the SPARQL query is captured to a
local variable, `query`, but the query is not passed on to the graph
store.

This results in a dead-end in the code, and causes the SPARQL client
which made the initial HTTP POST request to receive an
invalid/unexpected response from the graph store, given that the query
string was not actually submitted to the graph store for evaluation.

The fix for this issue was straight-forward and involved simply
creating a new dictionary from any possible additional key-values in
the `request.form` object and the SPARQL query captured to the local
`query` variable which is then be passed to the existing
`execute_query_post()`, method’s `data` argument, allowing the query to
be submitted to the graph store successfully.

This fix is backwards-compatible with the existing functionality of LOD
Gateway.

This commit also fixes a typo in the service name for Fuseki in the
`.env.example` file in both the `SPARQL_QUERY_ENDPOINT` and
`SPARQL_UPDATE_ENDPOINT` environment variables, where the service name
was misspelt which prevented LOD Gateway from communicating with the
Fuseki instance used in local development Docker setups of LOD Gateway.

Fixed a few minor spacing/grammatical issues in the `README` file and
incremented the copyright year.